DataList tree view for Images
This is an initial pass at extending the DataList component
to support a hierarchical structure. Some things are missing
or not fully flushed out (like showing status, showing a kebab
instead of the Download action, setting widths on content
cells to maintain alignment, including an icon per list item,
and other miscellaneous requirements).